ci: clear the clippy warnings and make the job blocking

+++ crates/zkml-prover/src/onnx/extract.rs
@@ -74,7 +74,7 @@ pub fn extract_linear_classifier(node: &NodeProto) -> Result<LogisticRegression,
 
     // Convert to f64 for quantization
     let weights_f64: Vec<f64> = coefficients.to_vec();
-    let bias_f64 = intercepts[0] as f64;
+    let bias_f64 = intercepts[0];
 
     // Quantize using existing helpers
     let quantized_weights = weights_f64
